rutil: Data: Fix out of bounds read
In case a data object contained content with a binary
zero in it and it was comared to a string literal that
matched the content before the binary zero, the
rhs[lhs.mSize] check could be out of bounds.

This issue was found while fuzzing the SDP parser
with LLVM libFuzzer.